third_party/blink/web_tests/fast/forms/tabs-with-modifiers-expected.txt

This test tabs with different combinations of modifiers pressed, where some should advance focus and some should not.

resetting focus to middle input
- focus event: b
test tab case that should advance focus
- keydown event: b
- blur event: b
- focus event: c
resetting focus to middle input
- blur event: c
- focus event: b
test tab case that should advance focus
- keydown event: b
- blur event: b
- focus event: a
resetting focus to middle input
- blur event: a
- focus event: b
test tab case that should not advance focus
- keydown event: b
resetting focus to middle input
test tab case that should not advance focus
- keydown event: b
resetting focus to middle input
test tab case that should not advance focus
- keydown event: b
resetting focus to middle input
test tab case that should not advance focus
- keydown event: b
resetting focus to middle input
test tab case that should not advance focus
- keydown event: b
resetting focus to middle input
test tab case that should not advance focus
- keydown event: b
